상품등록API: ProductImage 저장되지 않는 문제 수정
